Cleanup duplication across web and visual
